Fix OverflowException in sync state and improve HTTP handling
- Add validation for negative values in ServerInfo.UpdateFolderSyncState
- Add PruneStaleFolderStates method to clean up stale folder states
- Add diagnostic event for invalid sync data detection
- Fix Connected Devices display showing -1 when no devices configured
- Improve HTTP request handling in SyncthingEventsListener and SyncthingMonitorService using HttpRequestMessage
- Fix MudBlazor 8.x warnings (Dense, Title, IsInitiallyExpanded attributes)
- Fix obsolete WatchSystemPreference warning in App.razor
- Remove unnecessary System.Text.Json package from WebUI
- Fix uninitialized field warnings in Profiling, Statistics, NavMenu

+++ src/CreatioHelper.Application/Interfaces/ISyncEngine.cs
@@ -70,17 +70,17 @@ public class SyncStatus
     public long GlobalBytes { get; set; }
     public long LocalBytes { get; set; }
     public long NeedBytes { get; set; }
-    public int GlobalFiles { get; set; }
-    public int LocalFiles { get; set; }
-    public int NeedFiles { get; set; }
-    public int NeedDeletes { get; set; }
+    public long GlobalFiles { get; set; }
+    public long LocalFiles { get; set; }
+    public long NeedFiles { get; set; }
+    public long NeedDeletes { get; set; }
     public DateTime LastScan { get; set; }
     public DateTime LastSync { get; set; }
     public List<string> Errors { get; set; } = new();
     
     // Additional properties for Syncthing compatibility
-    public int TotalFiles { get; set; }
-    public int TotalDirectories { get; set; }
+    public long TotalFiles { get; set; }
+    public long TotalDirectories { get; set; }
     public long TotalBytes { get; set; }
     public int LocalDirectories { get; set; }
     public int OutOfSyncFiles { get; set; }
@@ -200,17 +200,17 @@ public class FolderCompletionStatus
     /// <summary>
     /// Total global items (files + directories)
     /// </summary>
-    public int GlobalItems { get; set; }
+    public long GlobalItems { get; set; }
 
     /// <summary>
     /// Items needed to complete sync
     /// </summary>
-    public int NeedItems { get; set; }
+    public long NeedItems { get; set; }
 
     /// <summary>
     /// Deletes needed
     /// </summary>
-    public int NeedDeletes { get; set; }
+    public long NeedDeletes { get; set; }
 
     /// <summary>
     /// Current sequence number
